backend/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
pypdf
openai
python-dotenv
numpy
